+++ src/renderer/class/Directory.js
@@ -107,9 +107,7 @@ export default class Directory {
    * @memberof Directory
    */
   selectByFile(file) {
-    //const index = this.getFileIndex(file);
-
-    const index = this.files.indexOf(file);
+    const index = this.getFileIndex(file);
 
     if (-1 === index) {
       return;