+++ src/renderer/components/JsonPreview.vue
@@ -1,14 +1,27 @@
 <template>
   <div class="json">
-    <v-jsoneditor ref="jsonEditor" class="json__editor" :options="options" :plus="false" height="400px" />
+    <v-jsoneditor ref="jsonEditor" class="json__editor" :options="options" :plus="false" height="446px" />
     <div class="json__button-group">
+      <v-dialog v-model="saveErrorDialog" width="500">
+        <v-card>
+          <v-card-title class="error" primary-title>
+            <h2 class="font-semi-bold json__dialog-title">Warning</h2>
+          </v-card-title>
+          <v-card-text>
+            <span class="font-16">There are errors in the caption JSON. It is recommended you correct those before saving, otherwise your changes to those specific captions will not be saved.</span>
+          </v-card-text>
+          <v-card-actions>
+            <v-spacer></v-spacer>
+            <v-btn color="accent" class="font-semi-bold font-16 --capital" @click="saveErrorDialog = false">
+              Cancel
+            </v-btn>
+            <v-btn color="error" class="font-semi-bold font-16 --capital" @click="onSave(null, null, true)">Save</v-btn>
+          </v-card-actions>
+        </v-card>
+      </v-dialog>
       <v-dialog v-model="dialog" width="500">
         <template v-slot:activator="{on}">
-          <v-btn
-            color="error"
-            class="font-semi-bold font-16 --capital json__button-cancel"
-            v-on="on"
-          >
+          <v-btn color="error" class="font-semi-bold font-16 --capital json__button-cancel" v-on="on">
             Clear
           </v-btn>
         </template>
@@ -17,29 +30,18 @@
             <h2 class="font-semi-bold json__dialog-title">Warning</h2>
           </v-card-title>
           <v-card-text>
-            <span class>This will clear all captions.</span>
+            <span class="font-16">This will clear all captions.</span>
           </v-card-text>
           <v-card-actions>
             <v-spacer></v-spacer>
-            <v-btn
-              color="accent"
-              class="font-semi-bold font-16 --capital"
-              @click="dialog = false"
-            >
+            <v-btn color="accent" class="font-semi-bold font-16 --capital" @click="dialog = false">
               Cancel
             </v-btn>
             <v-btn color="error" class="font-semi-bold font-16 --capital" @click="reset">Ok</v-btn>
           </v-card-actions>
         </v-card>
       </v-dialog>
-      <v-btn
-        download="export.json"
-        target="_blank"
-        :href="blob"
-        color="accent"
-        class="font-semi-bold font-16 --capital json__button-export"
-        :disabled="Object.keys(jsonErrors).length > 0"
-      >
+      <v-btn download="export.json" target="_blank" :href="blob" color="accent" class="font-semi-bold font-16 --capital json__button-export" :disabled="Object.keys(jsonErrors).length > 0">
         Export Code
       </v-btn>
     </div>
@@ -54,6 +56,12 @@
 <script>
 import { EventBus } from '@/renderer/class/EventBus';
 import VJsoneditor from 'v-jsoneditor';
+import { ipcRenderer, path } from 'electron';
+import { mapState } from 'vuex';
+import { EVENTS } from '../../contents';
+
+const fs = require('fs');
+
 export default {
   components: {
     VJsoneditor
@@ -69,6 +77,7 @@ export default {
       data,
       blob: null,
       dialog: false,
+      saveErrorDialog: false,
       jsonErrors: false,
       currentIndex: 0,
       origin: 'JsonPreview',
@@ -81,6 +90,23 @@ export default {
       },
     };
   },
+  computed: {
+    ...mapState({
+
+      /**
+       * Returns the path for the current project audio files
+       */
+      captionLocation: function (state) {
+        return state.captionInfo.captionLocation;
+      },
+      /**
+       * returns whether or not there are unsaved caption changes
+       */
+      isUnsavedChanges: function (state) {
+        return state.captionInfo.isUnsavedChanges;
+      }
+    })
+  },
   /**
    *
    */
@@ -89,6 +115,16 @@ export default {
     EventBus.$on('caption_update', this.onUpdate);
     EventBus.$on('caption_changed', this.onCaptionChange);
     EventBus.$on('caption_data', this.update);
+    EventBus.$on('file_list_generated', this.createFileNameMap);
+    EventBus.$on(EVENTS.SAVE_CAPTION_DATA, this.onSave);
+    ipcRenderer.on(EVENTS.SAVE_CAPTION_DATA, this.onSave);
+    ipcRenderer.on(EVENTS.CLEAR_CAPTION_DATA, this.onMenuClear);
+    ipcRenderer.on(EVENTS.OPEN_CAPTION_FILE, this.onCaptionFileOpen);
+
+    //If a caption file has been previously saved/opened load it in on caption studio start up
+    if (this.captionLocation) {
+      this.onCaptionFileOpen(null, this.captionLocation);
+    }
   },
   /**
    *
@@ -97,6 +133,10 @@ export default {
     EventBus.$off('caption_update', this.onUpdate);
     EventBus.$off('caption_data', this.update);
     EventBus.$off('caption_changed', this.onCaptionChange);
+    ipcRenderer.removeListener(EVENTS.SAVE_CAPTION_DATA, this.onSave);
+    ipcRenderer.removeListener(EVENTS.CLEAR_CAPTION_DATA, this.onMenuClear);
+    ipcRenderer.removeListener(EVENTS.OPEN_CAPTION_FILE, this.onCaptionFileOpen);
+    this.json = '';
   },
   methods: {
     /**
@@ -104,7 +144,48 @@ export default {
      */
     onEdit($event) {
       this.checkErrors($event, this.origin);
+      if (this.jsonErrors) {
+        console.log('errors?', this.jsonErrors);
+        return;
+      }
       EventBus.$emit('json_update', $event, this.origin);
+    },
+    /**
+     * Handles the save caption event from app menu or keyboard shortcut
+     */
+    onSave(event, filePath, force = false) {
+
+      if (!filePath) {
+        filePath = this.captionLocation;
+      }
+
+      if (this.jsonErrors && !force) {
+        this.saveErrorDialog = true;
+        return;
+      }
+      this.saveErrorDialog = false;
+
+      fs.writeFile(filePath, this.json, err => {
+        if (err) {
+          throw err;
+        }
+        this.$store.dispatch('setIsUnsavedChanges', { isUnsavedChanges: false });
+        console.log('JSON data is saved.');
+      });
+    },
+    /**
+     * Handles opening/loading the user provided caption file
+     */
+    onCaptionFileOpen(event, filePath) {
+      fs.readFile(filePath, 'utf-8', (err, data) => {
+        this.update(JSON.parse(data.toString()), 'userOpen');
+      });
+    },
+    /**
+     * Handles the clear event sent from the app menu
+     */
+    onMenuClear() {
+      this.dialog = true;
     },
     /**
      *
@@ -142,7 +223,11 @@ export default {
     /**
      *
      */
-    onCaptionChange({ index, file, name }) {
+    onCaptionChange({
+      index,
+      file,
+      name
+    }) {
       this.activeFile = name;
       this.fileNameMap[name] = file;
       this.currentIndex = index;
@@ -153,14 +238,28 @@ export default {
     update(data, $origin) {
       this.checkErrors(data, $origin);
 
-      if ($origin === this.origin) {
-        return;
-      }
-
       this.data = this.cleanData(data);
       this.$refs.jsonEditor.editor.update(this.data);
       this.json = JSON.stringify(this.data, null, 2);
       this.createBlob();
+
+      if ($origin === 'userOpen') {
+        this.checkErrors(JSON.parse(this.json), this.origin);
+        EventBus.$emit('json_update', JSON.parse(this.json), $origin);
+      }
+
+    },
+    /**
+     *
+     */
+    createFileNameMap($event) {
+      if (!Array.isArray($event)) {
+        return;
+      }
+
+      $event.forEach(file => {
+        this.fileNameMap[file.name.replace(/.(ogg|mp3|mpeg)$/, '').trim()] = file;
+      });
     },
     /**
      *
@@ -174,9 +273,13 @@ export default {
         }
 
         const reduced = data[key[i]].reduce((filtered, e) => {
-          if ( (e.content && e.start < e.end) ) {
+          if ((e.content && e.start < e.end)) {
             if (e.content.trim()) {
-              filtered.push({content: e.content.replace(/\n$/, ''), start: e.start, end: e.end});
+              filtered.push({
+                content: e.content.replace(/\n$/, ''),
+                start: e.start,
+                end: e.end
+              });
             }
           }
           return filtered;
@@ -193,7 +296,9 @@ export default {
      */
     createBlob() {
       this.blob = URL.createObjectURL(
-        new Blob([JSON.stringify(this.data)], { type: 'application/json' })
+        new Blob([JSON.stringify(this.data)], {
+          type: 'application/json'
+        })
       );
     },
     /**
@@ -205,7 +310,7 @@ export default {
       this.update({});
     },
     /**
-     *
+     * Ensure that each caption line has a value, and has proper time values
      */
     validateJSON(json, $origin) {
       const errors = {};
@@ -214,7 +319,7 @@ export default {
         const file = json[key];
         file.forEach((caption, index) => {
           if (caption.edited || $origin === this.origin) {
-            if (!caption.content || !caption.content.trim() ) {
+            if (!caption.content || !caption.content.trim()) {
               errors[key].push(`Error at caption [${key}], index [${index}]: Caption content must be non-empty`);
             }
             if ('number' !== typeof caption.start || caption.start < 0) {
@@ -284,7 +389,7 @@ $menu-height: 5.6rem;
     &-outer.has-main-menu-bar {
       margin-top: 0;
       padding-top: 0;
-      height: calc(100% - #{$menu-height});
+      height: 100%;
     }
   }
 
@@ -321,8 +426,7 @@ $menu-height: 5.6rem;
       align-items: center;
       border-radius: 2rem;
 
-
-      &> * {
+      &>* {
         width: 50%;
       }
     }